+++ flutter/test/mocks.dart
@@ -9,7 +9,7 @@ import 'package:sentry/src/platform/platform.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 
 import 'package:meta/meta.dart';
-import 'package:sentry_flutter/src/binding_utils.dart';
+import 'package:sentry_flutter/src/binding_wrapper.dart';
 import 'package:sentry_flutter/src/renderer/renderer.dart';
 import 'package:sentry_flutter/src/sentry_native.dart';
 import 'package:sentry_flutter/src/sentry_native_channel.dart';
